Extract prim fun again out of concerns of maintanability
One advantage of keeping prim funs not in the main expression definition
is that it separates out the core lambda calculus (with pattern matching
and let-bindings) from the language-specific aspects, meaning we may be
able to write passes that are only concerned with the expression
language without considering side effects.

I annotated each node of the AST with the type. Even though this
introduces redundancy (and thus opportunity for inconsistency), I wanted
to do so for consistency (it can become confusing what 't' refers to in
a data constructor), and to maintain flexibility on the "shape" of the
type system (for instance, the fact that a Lambda's type is always a ->
b is a concern of the type system rather than the expression). This
separation of concerns should help with the modularity of our compiler
architecture at this earlier stage, when some core definitions are still
in flux.